feat(data-exploration): convert funnel properties correlation table (#14992)
