+++ frontend/src/scenes/funnels/funnelLogic.test.ts
@@ -1,20 +1,11 @@
-import { DEFAULT_EXCLUDED_PERSON_PROPERTIES, funnelLogic } from './funnelLogic'
-import { MOCK_DEFAULT_TEAM, MOCK_TEAM_ID } from 'lib/api.mock'
-import { expectLogic, partial } from 'kea-test-utils'
+import { funnelLogic } from './funnelLogic'
+import { MOCK_TEAM_ID } from 'lib/api.mock'
+import { expectLogic } from 'kea-test-utils'
 import { initKeaTests } from '~/test/init'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { insightLogic } from 'scenes/insights/insightLogic'
-import {
-    AvailableFeature,
-    CorrelationConfigType,
-    FunnelCorrelationResultsType,
-    FunnelsFilterType,
-    FunnelVizType,
-    InsightLogicProps,
-    InsightShortId,
-    InsightType,
-} from '~/types'
+import { AvailableFeature, InsightLogicProps, InsightShortId, InsightType } from '~/types'
 import { teamLogic } from 'scenes/teamLogic'
 import { userLogic } from 'scenes/userLogic'
 import { router } from 'kea-router'
@@ -26,86 +17,7 @@ import { openPersonsModal } from 'scenes/trends/persons-modal/PersonsModal'
 
 jest.mock('scenes/trends/persons-modal/PersonsModal')
 
-const Insight12 = '12' as InsightShortId
 const Insight123 = '123' as InsightShortId
-
-export const mockInsight = {
-    id: Insight123,
-    short_id: 'SvoU2bMC',
-    name: null,
-    filters: {
-        breakdown: null,
-        breakdown_type: null,
-        display: 'FunnelViz',
-        events: [
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 0,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 1,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 2,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-            {
-                id: '$pageview',
-                type: 'events',
-                order: 3,
-                name: '$pageview',
-                custom_name: null,
-                math: null,
-                math_property: null,
-                properties: [],
-            },
-        ],
-        funnel_from_step: 0,
-        funnel_to_step: 1,
-        funnel_viz_type: 'steps',
-        insight: 'FUNNELS',
-        layout: 'vertical',
-    },
-    order: null,
-    deleted: false,
-    dashboard: null,
-    layouts: {},
-    color: null,
-    last_refresh: null,
-    result: null,
-    created_at: '2021-09-22T18:22:20.036153Z',
-    description: null,
-    updated_at: '2021-09-22T19:03:49.322258Z',
-    tags: [],
-    favorited: false,
-    saved: false,
-    created_by: {
-        id: 1,
-        uuid: '017c0441-bcb2-0000-bccf-dfc24328c5f3',
-        distinct_id: 'fM7b6ZFi8MOssbkDI55ot8tMY2hkzrHdRy1qERa6rCK',
-        first_name: 'Alex',
-        email: 'alex@posthog.com',
-    },
-}
 
 const funnelResults = [
     {
@@ -133,200 +45,24 @@ const funnelResults = [
 
 describe('funnelLogic', () => {
     let logic: ReturnType<typeof funnelLogic.build>
-    let correlationConfig: CorrelationConfigType = {}
 
     beforeEach(() => {
         useAvailableFeatures([AvailableFeature.CORRELATION_ANALYSIS, AvailableFeature.GROUP_ANALYTICS])
         useMocks({
             get: {
-                '/api/projects/@current': () => [
-                    200,
-                    {
-                        ...MOCK_DEFAULT_TEAM,
-                        correlation_config: correlationConfig,
-                    },
-                ],
-                '/api/projects/:team/insights/': (req) => {
-                    if (req.url.searchParams.get('saved')) {
-                        return [
-                            200,
-                            {
-                                results: funnelResults,
-                            },
-                        ]
-                    }
-                    const shortId = req.url.searchParams.get('short_id') || ''
-                    if (shortId === '500') {
-                        return [500, { status: 0, detail: 'error from the API' }]
-                    }
-                    return [
-                        200,
-                        {
-                            results: [mockInsight],
-                        },
-                    ]
+                '/api/projects/:team/insights/': {
+                    results: [{}],
                 },
-                '/api/projects/:team/insights/trend/': { results: ['trends result from api'] },
+                '/api/projects/:team/insights/:id/': {},
                 '/api/projects/:team/groups_types/': [],
-                '/some/people/url': { results: [{ people: [] }] },
-                '/api/projects/:team/persons/funnel': { results: [], next: null },
-                '/api/projects/:team/persons/properties': [
-                    { name: 'some property', count: 20 },
-                    { name: 'another property', count: 10 },
-                    { name: 'third property', count: 5 },
-                ],
-                '/api/projects/:team/groups/property_definitions': {
-                    '0': [
-                        { name: 'industry', count: 2 },
-                        { name: 'name', count: 1 },
-                    ],
-                    '1': [{ name: 'name', count: 1 }],
-                },
-            },
-            patch: {
-                '/api/projects/:id': (req) => [
-                    200,
-                    {
-                        ...MOCK_DEFAULT_TEAM,
-                        correlation_config: {
-                            ...correlationConfig,
-                            excluded_person_property_names: (req.body as any)?.correlation_config
-                                ?.excluded_person_property_names,
-                        },
-                    },
-                ],
             },
             post: {
-                '/api/projects/:team/insights/': (req) => [
-                    200,
-                    { id: 12, short_id: Insight12, ...((req.body as any) || {}) },
-                ],
-                '/api/projects/:team/insights/:id/viewed': [201],
                 '/api/projects/:team/insights/funnel/': {
-                    is_cached: true,
-                    last_refresh: '2021-09-16T13:41:41.297295Z',
                     result: funnelResults,
-                    type: 'Funnel',
-                },
-                '/api/projects/:team/insights/funnel/correlation': (req) => {
-                    const data = req.body as any
-                    if (data?.funnel_correlation_type === 'properties') {
-                        const excludePropertyFromProjectNames = data?.funnel_correlation_exclude_names || []
-                        const includePropertyNames = data?.funnel_correlation_names || []
-                        return [
-                            200,
-                            {
-                                is_cached: true,
-                                last_refresh: '2021-09-16T13:41:41.297295Z',
-                                result: {
-                                    events: [
-                                        {
-                                            event: { event: 'some property' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'success',
-                                        },
-                                        {
-                                            event: { event: 'another property' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'failure',
-                                        },
-                                    ]
-                                        .filter(
-                                            (correlation) =>
-                                                includePropertyNames.includes('$all') ||
-                                                includePropertyNames.includes(correlation.event.event)
-                                        )
-                                        .filter(
-                                            (correlation) =>
-                                                !excludePropertyFromProjectNames.includes(correlation.event.event)
-                                        ),
-                                },
-                                type: 'Funnel',
-                            },
-                        ]
-                    } else if (data?.funnel_correlation_type === 'events') {
-                        return [
-                            200,
-                            {
-                                is_cached: true,
-                                last_refresh: '2021-09-16T13:41:41.297295Z',
-                                result: {
-                                    events: [
-                                        {
-                                            event: { event: 'some event' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'success',
-                                        },
-                                        {
-                                            event: { event: 'another event' },
-                                            success_count: 1,
-                                            failure_count: 1,
-                                            odds_ratio: 1,
-                                            correlation_type: 'failure',
-                                        },
-                                    ],
-                                },
-                                type: 'Funnel',
-                            },
-                        ]
-                    } else if (data?.funnel_correlation_type === 'event_with_properties') {
-                        const targetEvent = data?.funnel_correlation_event_names[0]
-                        const excludedProperties = data?.funnel_correlation_event_exclude_property_names
-                        return [
-                            200,
-                            {
-                                result: {
-                                    events: [
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 29,
-                                            correlation_type: 'success',
-                                            event: { event: `some event::name::Hester` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 29,
-                                            correlation_type: 'success',
-                                            event: { event: `some event::Another name::Alice` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 25,
-                                            correlation_type: 'success',
-                                            event: { event: `another event::name::Aloha` },
-                                        },
-                                        {
-                                            success_count: 1,
-                                            failure_count: 0,
-                                            odds_ratio: 25,
-                                            correlation_type: 'success',
-                                            event: { event: `another event::Another name::Bob` },
-                                        },
-                                    ].filter(
-                                        (record) =>
-                                            record.event.event.split('::')[0] === targetEvent &&
-                                            !excludedProperties.includes(record.event.event.split('::')[1])
-                                    ),
-                                    last_refresh: '2021-11-05T09:26:16.175923Z',
-                                    is_cached: false,
-                                },
-                            },
-                        ]
-                    }
                 },
             },
         })
         initKeaTests(false)
-        window.POSTHOG_APP_CONTEXT = undefined // to force API request to /api/project/@current
     })
 
     const defaultProps: InsightLogicProps = {
@@ -616,157 +352,6 @@ describe('funnelLogic', () => {
                 title: expect.any(Object),
                 url: '/some/people/url?funnel_step=2&funnel_step_breakdown=Latvia', // Series funnel_step_breakdown included
             })
-        })
-    })
-
-    describe('funnel correlation properties', () => {
-        const props = { dashboardItemId: Insight123, syncWithUrl: true }
-
-        it('Selecting all properties returns expected result', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => logic.actions.setPropertyNames(logic.values.allProperties))
-                .toFinishListeners()
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [
-                            {
-                                event: { event: 'some property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'success',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                            {
-                                event: { event: 'another property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'failure',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                        ],
-                    },
-                })
-        })
-
-        it('Deselecting all returns empty result', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => logic.actions.setPropertyNames([]))
-                .toDispatchActions(logic, ['loadPropertyCorrelationsSuccess'])
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [],
-                    },
-                })
-        })
-
-        it('are not triggered when results are loaded, when trends visualisation set', async () => {
-            await initFunnelLogic(props)
-            await expectLogic(logic, () => {
-                logic.actions.loadResultsSuccess({
-                    filters: {
-                        insight: InsightType.FUNNELS,
-                        funnel_viz_type: FunnelVizType.Trends,
-                    } as FunnelsFilterType,
-                })
-            }).toNotHaveDispatchedActions(['loadEventCorrelations', 'loadPropertyCorrelations'])
-        })
-
-        it('triggers update to correlation list when property excluded from project', async () => {
-            userLogic.mount()
-            await initFunnelLogic(props)
-            //
-            // // Make sure we have loaded the team already
-            // await expectLogic(teamLogic, () => teamLogic.actions.loadCurrentTeam()).toFinishAllListeners()
-
-            await expectLogic(logic, () => {
-                logic.actions.setPropertyNames(logic.values.allProperties)
-                logic.actions.loadResultsSuccess({ filters: { insight: InsightType.FUNNELS } })
-                logic.actions.excludePropertyFromProject('another property')
-            })
-                .toFinishAllListeners()
-                .toMatchValues({
-                    propertyNames: ['some property', 'third property'],
-                    excludedPropertyNames: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['another property']),
-                    allProperties: ['some property', 'third property'],
-                })
-
-            expect(logic.values.propertyCorrelationValues).toEqual([
-                {
-                    event: { event: 'some property' },
-                    success_count: 1,
-                    failure_count: 1,
-                    odds_ratio: 1,
-                    correlation_type: 'success',
-                    result_type: FunnelCorrelationResultsType.Properties,
-                },
-            ])
-        })
-
-        it('isPropertyExcludedFromProject returns true initially, then false when excluded, and is persisted to team config', async () => {
-            await initFunnelLogic(props)
-
-            expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(false)
-
-            await expectLogic(logic, () =>
-                logic.actions.excludePropertyFromProject('some property')
-            ).toFinishAllListeners()
-
-            expect(logic.values.isPropertyExcludedFromProject('some property')).toBe(true)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: {
-                        excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
-                    },
-                }),
-            })
-
-            // Also make sure that excluding the property again doesn't double
-            // up on the config list
-            await expectLogic(logic, () =>
-                logic.actions.excludePropertyFromProject('some property')
-            ).toFinishAllListeners()
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: {
-                        excluded_person_property_names: DEFAULT_EXCLUDED_PERSON_PROPERTIES.concat(['some property']),
-                    },
-                }),
-            })
-        })
-
-        it('loads property exclude list from Project settings', async () => {
-            correlationConfig = { excluded_person_property_names: ['some property'] }
-            await initFunnelLogic(props)
-
-            await expectLogic(teamLogic).toMatchValues({
-                currentTeam: partial({
-                    correlation_config: { excluded_person_property_names: ['some property'] },
-                }),
-            })
-
-            await expectLogic(logic, () => {
-                logic.actions.setPropertyNames(logic.values.allProperties)
-                logic.actions.loadResultsSuccess({ filters: { insight: InsightType.FUNNELS } })
-            })
-                .toFinishAllListeners()
-                .toMatchValues({
-                    propertyCorrelations: {
-                        events: [
-                            {
-                                event: { event: 'another property' },
-                                success_count: 1,
-                                failure_count: 1,
-                                odds_ratio: 1,
-                                correlation_type: 'failure',
-                                result_type: FunnelCorrelationResultsType.Properties,
-                            },
-                        ],
-                    },
-                })
         })
     })
